add TokenUpgraded
