+++ packages/subgraph/src/mappings/superToken.ts
@@ -1,33 +1,35 @@
-import {
-    BigDecimal,
-    BigInt,
-    EthereumEvent,
-    log,
-} from "@graphprotocol/graph-ts";
-
-import { SuperTokenCreated as SuperTokenCreatedEvent } from "../../generated/SuperTokenFactory/ISuperTokenFactory";
+import { dataSource } from "@graphprotocol/graph-ts";
 
 import {
-    Transaction,
-    Account,
-    Token,
-    SuperTokenCreated,
-} from "../../generated/schema";
+    // ISuperToken as SuperToken,
+    TokenUpgraded as TokenUpgradedEvent,
+    TokenDowngraded as TokenDowngradedEvent,
+} from "../../generated/templates/SuperToken/ISuperToken";
+
+import { TokenUpgraded, TokenDowngraded } from "../../generated/schema";
 
 import { createEventID, logTransaction, fetchToken } from "../utils";
 
-export function handleSuperTokenCreated(event: SuperTokenCreatedEvent): void {
-    let address = event.params.token.toHex();
-    let token = fetchToken(address);
-    token.save();
+export function handleTokenUpgraded(event: TokenUpgradedEvent): void {
+    let account = event.params.account.toHex();
+    let amount = event.params.amount;
 
-    // TODO: start monitoring events for this token?
-
-    let ev = new SuperTokenCreated(createEventID(event));
+    let ev = new TokenUpgraded(createEventID(event));
     ev.transaction = logTransaction(event).id;
-    ev.address = event.params.token;
-    ev.underlyingAddress = token.underlyingAddress;
-    ev.symbol = token.symbol;
-    ev.name = token.name;
+    ev.account = account;
+    ev.amount = amount;
+    ev.token = dataSource.address().toHex();
+    ev.save();
+}
+
+export function handleTokenDowngraded(event: TokenDowngradedEvent): void {
+    let account = event.params.account.toHex();
+    let amount = event.params.amount;
+
+    let ev = new TokenDowngraded(createEventID(event));
+    ev.transaction = logTransaction(event).id;
+    ev.account = account;
+    ev.amount = amount;
+    ev.token = dataSource.address().toHex();
     ev.save();
 }
